requirements.txt
Flask==3.0.0
Flask_Login==0.6.2
Flask_SocketIO==5.3.6
flask_sqlalchemy==3.1.1
flask_wtf==1.2.1
Pillow==10.1.0
Pillow==10.1.0
Werkzeug==2.2.3
WTForms==3.1.0
